refactor: split order and cancel physical card into separate modals
- OrderPhysicalCardModal now only handles the shipping form and order
- New CancelPhysicalCardModal handles the cancel flow with Dialog
  confirmation following DiscardChangesDialog pattern
- details.tsx opens the correct modal based on hasPhysicalCard state
- Renamed props to onPhysicalCardPress for clarity
